requirements.txt
pandas==2.2.1
numpy==1.26.4
usgs-libcomcat==2.0.23
matplotlib==3.8.3
pyspark==3.5.1
catboost==1.2.3
scikit-learn==1.4.1.post1
